Add support for file thumbnails.
File thumbnail URLs need to be fetched with an Authorization header, so we can't
return them as is. Instead we return an encrypted (account ID, file ID) tuple
embedded into one of our URLs. When fetched, we can decode the tuple, look up
the account (and its API token), fetch the thumbnail with the header and proxy
back the response.