Increase default task queue rate a bit.
We end up generating a separate task for each channel (regardless of whether it
has any messages or not), so it was taking a while to go through all of the
digests (mine were being sent out as laste as 6am). Let's see if this higher
rate is still enough to stay under rate limits.